+++ mod-source-record-manager-server/src/main/java/org/folio/services/ChangeEngineServiceImpl.java
@@ -2,22 +2,28 @@ package org.folio.services;
 
 import io.vertx.core.Future;
 import io.vertx.core.Promise;
-import io.vertx.core.http.HttpClientResponse;
+import io.vertx.core.Vertx;
+import io.vertx.core.json.Json;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
+import io.vertx.kafka.client.producer.KafkaHeader;
+import io.vertx.kafka.client.producer.KafkaProducer;
+import io.vertx.kafka.client.producer.KafkaProducerRecord;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.mutable.MutableInt;
-import org.folio.HttpStatus;
 import org.folio.dao.JobExecutionSourceChunkDao;
 import org.folio.dataimport.util.OkapiConnectionParams;
-import org.folio.dataimport.util.Try;
-import org.folio.rest.client.SourceStorageBatchClient;
+import org.folio.kafka.KafkaConfig;
+import org.folio.kafka.KafkaTopicNameHelper;
+import org.folio.processing.events.utils.ZIPArchiver;
 import org.folio.rest.jaxrs.model.ActionProfile;
 import org.folio.rest.jaxrs.model.ActionProfile.Action;
 import org.folio.rest.jaxrs.model.ActionProfile.FolioRecord;
 import org.folio.rest.jaxrs.model.ErrorRecord;
+import org.folio.rest.jaxrs.model.Event;
+import org.folio.rest.jaxrs.model.EventMetadata;
 import org.folio.rest.jaxrs.model.InitialRecord;
 import org.folio.rest.jaxrs.model.JobExecution;
 import org.folio.rest.jaxrs.model.JobExecutionSourceChunk;
@@ -28,7 +34,6 @@ import org.folio.rest.jaxrs.model.RawRecord;
 import org.folio.rest.jaxrs.model.RawRecordsDto;
 import org.folio.rest.jaxrs.model.Record;
 import org.folio.rest.jaxrs.model.RecordCollection;
-import org.folio.rest.jaxrs.model.RecordsBatchResponse;
 import org.folio.rest.jaxrs.model.RecordsMetadata;
 import org.folio.rest.jaxrs.model.StatusDto;
 import org.folio.services.afterprocessing.HrIdFieldService;
@@ -36,11 +41,13 @@ import org.folio.services.journal.JournalService;
 import org.folio.services.parsers.ParsedResult;
 import org.folio.services.parsers.RecordParser;
 import org.folio.services.parsers.RecordParserBuilder;
+import org.folio.util.pubsub.PubSubClientUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 import javax.ws.rs.NotFoundException;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Date;
@@ -48,38 +55,44 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
 import static java.lang.String.format;
-import static org.folio.HttpStatus.HTTP_CREATED;
 import static org.folio.rest.RestVerticle.MODULE_SPECIFIC_ARGS;
 import static org.folio.rest.jaxrs.model.JournalRecord.ActionType.CREATE;
 import static org.folio.services.afterprocessing.AdditionalFieldsUtil.TAG_999;
 import static org.folio.services.afterprocessing.AdditionalFieldsUtil.addFieldToMarcRecord;
 import static org.folio.services.afterprocessing.AdditionalFieldsUtil.getValue;
+import static org.folio.rest.jaxrs.model.DataImportEventTypes.DI_RAW_MARC_BIB_RECORDS_CHUNK_PARSED;
 
 @Service
 public class ChangeEngineServiceImpl implements ChangeEngineService {
 
-  private static final String CAN_T_CREATE_NEW_RECORDS_MSG = "Can't create new records with JobExecution id: %s in source-record-storage, response code %s";
   private static final Logger LOGGER = LoggerFactory.getLogger(ChangeEngineServiceImpl.class);
   private static final int THRESHOLD_CHUNK_SIZE =
     Integer.parseInt(MODULE_SPECIFIC_ARGS.getOrDefault("chunk.processing.threshold.chunk.size", "100"));
-  private static final String CAN_NOT_RETRIEVE_A_RESPONSE_MSG = "Can not retrieve a response. Reason is: %s";
 
   private JobExecutionSourceChunkDao jobExecutionSourceChunkDao;
   private JobExecutionService jobExecutionService;
   private JournalService journalService;
   private HrIdFieldService hrIdFieldService;
+  private KafkaConfig kafkaConfig;
+
+  //TODO: make it configurable
+  private int maxDistributionNum = 100;
+  private static final AtomicInteger indexer = new AtomicInteger();
 
   public ChangeEngineServiceImpl(@Autowired JobExecutionSourceChunkDao jobExecutionSourceChunkDao,
                                  @Autowired JobExecutionService jobExecutionService,
                                  @Autowired HrIdFieldService hrIdFieldService,
-                                 @Autowired @Qualifier("journalServiceProxy") JournalService journalService) {
+                                 @Autowired @Qualifier("journalServiceProxy") JournalService journalService,
+                                 @Autowired KafkaConfig kafkaConfig) {
     this.jobExecutionSourceChunkDao = jobExecutionSourceChunkDao;
     this.jobExecutionService = jobExecutionService;
     this.journalService = journalService;
     this.hrIdFieldService = hrIdFieldService;
+    this.kafkaConfig = kafkaConfig;
   }
 
   @Override
@@ -223,39 +236,67 @@ public class ChangeEngineServiceImpl implements ChangeEngineService {
       return Future.succeededFuture();
     }
 
-    return Try.itDo((Promise<List<Record>> promise) -> {
-      SourceStorageBatchClient client = new SourceStorageBatchClient(params.getOkapiUrl(), params.getTenantId(), params.getToken());
+    RecordCollection recordCollection = new RecordCollection()
+      .withRecords(parsedRecords)
+      .withTotalRecords(parsedRecords.size());
 
-      RecordCollection recordCollection = new RecordCollection()
-        .withRecords(parsedRecords)
-        .withTotalRecords(parsedRecords.size());
+    //TODO: this could be some utility method
+    Event event;
+    try {
+      event = new Event()
+        .withId(UUID.randomUUID().toString())
+        .withEventType(DI_RAW_MARC_BIB_RECORDS_CHUNK_PARSED.value())
+        .withEventPayload(ZIPArchiver.zip(Json.encode(recordCollection)))
+        .withEventMetadata(new EventMetadata()
+          .withTenantId(params.getTenantId())
+          .withEventTTL(1)
+          .withPublishedBy(PubSubClientUtils.constructModuleName()));
+    } catch (IOException e) {
+      e.printStackTrace();
+      LOGGER.error(e);
+      return Future.failedFuture(e);
+    }
 
-      client.postSourceStorageBatchRecords(recordCollection, response -> {
-        if (isStatus(response, HTTP_CREATED)) {
-          response.bodyHandler(it ->
-            promise.handle(
-              Try.itGet(() -> {
-                List<Record> createdRecords = it.toJsonObject().mapTo(RecordsBatchResponse.class).getRecords();
-                List<JsonObject> journalRecords = buildJournalRecordsForProcessedRecords(parsedRecords, createdRecords, CREATE);
-                journalService.saveBatch(new JsonArray(journalRecords), params.getTenantId());
-                return createdRecords;
-              })
-                .recover(ex -> Future.failedFuture(format(CAN_NOT_RETRIEVE_A_RESPONSE_MSG, ex.getMessage())))
-            )
-          );
-        } else {
-          List<JsonObject> journalRecords = buildJournalRecordsForProcessedRecords(parsedRecords, Collections.emptyList(), CREATE);
-          journalService.saveBatch(new JsonArray(journalRecords), params.getTenantId());
-          String message = format(CAN_T_CREATE_NEW_RECORDS_MSG, jobExecution.getId(), response.statusCode());
-          LOGGER.error(message);
-          promise.fail(message);
-        }
-      });
-    }).recover(e -> Future.failedFuture(format("Error during POST new records: %s", e.getMessage())));
-  }
+    String key = String.valueOf(indexer.incrementAndGet() % maxDistributionNum);
 
-  public static boolean isStatus(HttpClientResponse response, HttpStatus status) {
-    return response.statusCode() == status.toInt();
+    String topicName = KafkaTopicNameHelper.formatTopicName(kafkaConfig.getEnvId(), KafkaTopicNameHelper.getDefaultNameSpace(),
+      params.getTenantId(), DI_RAW_MARC_BIB_RECORDS_CHUNK_PARSED.value());
+
+    KafkaProducerRecord<String, String> record =
+      KafkaProducerRecord.create(topicName, key, Json.encode(event));
+
+    List<KafkaHeader> kafkaHeaders = params
+      .getHeaders()
+      .entries()
+      .stream()
+      .map(e -> KafkaHeader.header(e.getKey(), e.getValue()))
+      .collect(Collectors.toList());
+
+    record.addHeaders(kafkaHeaders);
+    record.addHeader("jobExecutionId", jobExecution.getId());
+
+    Promise<List<Record>> writePromise = Promise.promise();
+
+    String producerName = DI_RAW_MARC_BIB_RECORDS_CHUNK_PARSED + "_Producer";
+    KafkaProducer<String, String> producer =
+      KafkaProducer.createShared(Vertx.currentContext().owner(), producerName, kafkaConfig.getProducerProps());
+
+    producer.write(record, war -> {
+      producer.end(ear -> producer.close());
+      if (war.succeeded()) {
+        //TODO: this logic must be rewritten
+        List<JsonObject> journalRecords = buildJournalRecordsForProcessedRecords(parsedRecords, parsedRecords, CREATE);
+        journalService.saveBatch(new JsonArray(journalRecords), params.getTenantId());
+
+        writePromise.complete();
+      } else {
+        Throwable cause = war.cause();
+        LOGGER.error("{} write error:", cause, producerName);
+        writePromise.fail(cause);
+      }
+    });
+
+    return writePromise.future();
   }
 
   /**